renamed DAO package and added commented dependency to pom.xml
