+++ FreelanceExchange/src/main/java/org/freelance/DAO/AbstractHibernateDao.java
@@ -1,4 +1,4 @@
-package org.freelance.DAOs;
+package org.freelance.DAO;
 
 import java.util.List;
