+++ FreelanceExchange/src/main/java/org/freelance/DAO/RoleDao.java
@@ -1,9 +1,8 @@
-package org.freelance.DAOs;
+package org.freelance.DAO;
 
 import com.google.common.base.Preconditions;
 import jakarta.transaction.Transactional;
 import org.freelance.models.Role;
-import org.freelance.models.User;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
@@ -14,53 +13,30 @@ import java.util.List;
 
 @Repository
 @Transactional
-public class UserDao implements AbstractHibernateDao<User> {
+public class RoleDao implements AbstractHibernateDao<Role> {
     @Autowired
     private SessionFactory sessionFactory;
 
     @Override
-    public User find(long id) {
+    public Role find(long id) {
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
-        User response = session.get(User.class, id);
-        transaction.commit();
-        return response;
-    }
-
-    public User find(String login) {
-        Session session = sessionFactory.getCurrentSession();
-        Transaction transaction = session.beginTransaction();
-        User response = session.createQuery(
-                "SELECT u FROM User u WHERE u.login = '" + login + "'", User.class
-        ).getSingleResult();
+        Role response = session.get(Role.class, id);
         transaction.commit();
         return response;
     }
 
     @Override
-    public List<User> findAll() {
+    public List<Role> findAll() {
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
-        List<User> response = session.createQuery("SELECT u FROM User u", User.class).getResultList();
+        List<Role> response = session.createQuery("SELECT r FROM Role r", Role.class).getResultList();
         transaction.commit();
         return response;
     }
 
-//    public List<User> findByRole(String role) {
-//        Session session = sessionFactory.getCurrentSession();
-//        Transaction transaction = session.beginTransaction();
-//
-//        String query = "SELECT u FROM User u WHERE u.role_id = (SELECT r.id FROM Role r WHERE r.name = '" +
-//                role.toLowerCase() +
-//                "')";
-//
-//        List<User> response = session.createQuery(query, User.class).getResultList();
-//        transaction.commit();
-//        return response;
-//    }
-
     @Override
-    public User create(User entity) {
+    public Role create(Role entity) {
         Preconditions.checkNotNull(entity);
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
@@ -70,7 +46,7 @@ public class UserDao implements AbstractHibernateDao<User> {
     }
 
     @Override
-    public void update(User entity) {
+    public void update(Role entity) {
         Preconditions.checkNotNull(entity);
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
@@ -79,7 +55,7 @@ public class UserDao implements AbstractHibernateDao<User> {
     }
 
     @Override
-    public void delete(User entity) {
+    public void delete(Role entity) {
         Preconditions.checkNotNull(entity);
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
@@ -89,7 +65,7 @@ public class UserDao implements AbstractHibernateDao<User> {
 
     @Override
     public void delete(long id) {
-        User entity = find(id);
+        Role entity = find(id);
         delete(entity);
     }
 }
