+++ FreelanceExchange/src/main/java/org/freelance/DAO/UserDao.java
@@ -1,8 +1,8 @@
-package org.freelance.DAOs;
+package org.freelance.DAO;
 
 import com.google.common.base.Preconditions;
 import jakarta.transaction.Transactional;
-import org.freelance.models.Role;
+import org.freelance.models.User;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
 import org.hibernate.Transaction;
@@ -13,30 +13,53 @@ import java.util.List;
 
 @Repository
 @Transactional
-public class RoleDao implements AbstractHibernateDao<Role> {
+public class UserDao implements AbstractHibernateDao<User> {
     @Autowired
     private SessionFactory sessionFactory;
 
     @Override
-    public Role find(long id) {
+    public User find(long id) {
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
-        Role response = session.get(Role.class, id);
+        User response = session.get(User.class, id);
+        transaction.commit();
+        return response;
+    }
+
+    public User find(String login) {
+        Session session = sessionFactory.getCurrentSession();
+        Transaction transaction = session.beginTransaction();
+        User response = session.createQuery(
+                "SELECT u FROM User u WHERE u.login = '" + login + "'", User.class
+        ).getSingleResult();
         transaction.commit();
         return response;
     }
 
     @Override
-    public List<Role> findAll() {
+    public List<User> findAll() {
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
-        List<Role> response = session.createQuery("SELECT r FROM Role r", Role.class).getResultList();
+        List<User> response = session.createQuery("SELECT u FROM User u", User.class).getResultList();
         transaction.commit();
         return response;
     }
 
+//    public List<User> findByRole(String role) {
+//        Session session = sessionFactory.getCurrentSession();
+//        Transaction transaction = session.beginTransaction();
+//
+//        String query = "SELECT u FROM User u WHERE u.role_id = (SELECT r.id FROM Role r WHERE r.name = '" +
+//                role.toLowerCase() +
+//                "')";
+//
+//        List<User> response = session.createQuery(query, User.class).getResultList();
+//        transaction.commit();
+//        return response;
+//    }
+
     @Override
-    public Role create(Role entity) {
+    public User create(User entity) {
         Preconditions.checkNotNull(entity);
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
@@ -46,7 +69,7 @@ public class RoleDao implements AbstractHibernateDao<Role> {
     }
 
     @Override
-    public void update(Role entity) {
+    public void update(User entity) {
         Preconditions.checkNotNull(entity);
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
@@ -55,7 +78,7 @@ public class RoleDao implements AbstractHibernateDao<Role> {
     }
 
     @Override
-    public void delete(Role entity) {
+    public void delete(User entity) {
         Preconditions.checkNotNull(entity);
         Session session = sessionFactory.getCurrentSession();
         Transaction transaction = session.beginTransaction();
@@ -65,7 +88,7 @@ public class RoleDao implements AbstractHibernateDao<Role> {
 
     @Override
     public void delete(long id) {
-        Role entity = find(id);
+        User entity = find(id);
         delete(entity);
     }
 }
