+++ FreelanceExchange/src/main/java/org/freelance/services/RoleService.java
@@ -1,6 +1,6 @@
 package org.freelance.services;
 
-import org.freelance.DAOs.RoleDao;
+import org.freelance.DAO.RoleDao;
 import org.freelance.models.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
