+++ FreelanceExchange/src/main/java/org/freelance/services/TaskService.java
@@ -1,6 +1,6 @@
 package org.freelance.services;
 
-import org.freelance.DAOs.TaskDao;
+import org.freelance.DAO.TaskDao;
 import org.freelance.models.Task;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
